edgeless_node: move the creation of a node with metrics-collector resource from edgeless_orc to edgeless_node to reduce inter-crate dependencies and better readability
