+++ edgeless_node/src/resources/mod.rs
@@ -5,5 +5,6 @@ pub mod file_log;
 pub mod http_egress;
 pub mod http_ingress;
 pub mod kafka_egress;
+pub mod metrics_collector;
 pub mod ollama;
 pub mod redis;
